requirements.txt
fastapi
uvicorn
requests
python-dotenv
kafka-python